requirements.txt
boto3
pytest
moto
pillow